requirements.txt
asgiref==3.7.2
Django==5.0.2
django-appconf==1.0.6
django-ckeditor==6.7.1
django-debug-toolbar==4.3.0
django-filter==24.2
django-imagekit==5.0.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-phonenumber-field==7.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
Faker==24.4.0
phonenumberslite==8.13.34
pilkit==3.0
pillow==10.2.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4